Add files via upload
The best hyperparameters found during the search are:

Dropout rate: 0.3832
Learning rate: 0.0002156
LSTM units: 125.1

During the hyperparameter search, tested various combinations of dropout rate, learning rate, and the number of LSTM units in the model.

The table shows the results of different model configurations during the hyperparameter search. Each row in the table represents a different configuration, with columns indicating the iteration number, the target metric (accuracy in this case), and the values for dropout rate, learning rate, and LSTM units.